Initial commit
This initial commit creates a new Ansible Role for InfluxDB. This
includes everything needed to set up and install InfluxDB on a new host,
although it does not include any configuration changes or modifications.
This is enough to get the role ready for publishing.

All original work where possible licensed under the Mozilla Public
License 2.0.

Signed-off-by: Justin W. Flory <git@jwf.io>